requirements.txt
python-dotenv
chromadb==0.5.3
sentence-transformers
requests
PyPDF2
